create library.py, sort imports, and finish reformat oauth2.py
