+++ rblxopencloud/creator.py
@@ -1,15 +1,40 @@
-from .exceptions import InvalidAsset, ModeratedText, rblx_opencloudException
-import json, io
-from typing import Union, Optional, Iterable, TYPE_CHECKING
-import urllib3
-from enum import Enum
+# MIT License
+
+# Copyright (c) 2022-2024 treeben77
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
 from datetime import datetime
-from . import send_request, Operation
+from enum import Enum
+import io
+import json
+from typing import Iterable, Optional, TYPE_CHECKING, Union
+import urllib3
+
 from dateutil import parser
 
+from . import Operation, send_request
+from .exceptions import InvalidAsset, ModeratedText, rblx_opencloudException
+
 if TYPE_CHECKING:
-    from .user import User
     from .group import Group
+    from .user import User
 
 __all__ = (
     "AssetType",
