+++ rblxopencloud/datastore.py
@@ -1,9 +1,34 @@
-from .exceptions import rblx_opencloudException, NotFound, PreconditionFailed
-import json, datetime
-import base64, hashlib, urllib.parse
-from typing import Union, Optional, Iterable, TYPE_CHECKING
-from . import send_request, iterate_request
+# MIT License
+
+# Copyright (c) 2022-2024 treeben77
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+import datetime
+import json
+from typing import Iterable, Optional, TYPE_CHECKING, Union
+import urllib.parse
+
 from dateutil import parser
+
+from . import send_request, iterate_request
+from .exceptions import rblx_opencloudException, NotFound, PreconditionFailed
 
 if TYPE_CHECKING:
     from .experience import Experience
