+++ rblxopencloud/exceptions.py
@@ -1,4 +1,4 @@
-from typing import Union, Optional, TYPE_CHECKING
+from typing import Optional, TYPE_CHECKING, Union
 
 if TYPE_CHECKING:
     from .datastore import EntryInfo
@@ -25,15 +25,18 @@ class InvalidKey(rblx_opencloudException): pass
 class PermissionDenied(InvalidKey): pass
 class RateLimited(rblx_opencloudException): pass
 class ServiceUnavailable(rblx_opencloudException): pass
+
 class PreconditionFailed(rblx_opencloudException):
     def __init__(self, value, info, *args: object) -> None:
         self.value: Optional[Union[str, dict, list, int, float]] = value
         self.info: Optional[EntryInfo] = info
         super().__init__(*args)
+
 class InsufficientScope(InvalidKey):
     def __init__(self, scope, *args: object) -> None:
         self.required_scope: str = scope
         super().__init__(*args)
+
 class InvalidCode(InvalidKey): pass
 class InvalidAsset(rblx_opencloudException): pass
 class ModeratedText(rblx_opencloudException): pass
